+++ src/shared/tdoc-commons/services/tdoc-itemsjs.adapter.ts
@@ -1,4 +1,4 @@
-import {TourDocRecord, TourDocRecordRelation} from '../model/records/tdoc-record';
+import {TourDocRecord} from '../model/records/tdoc-record';
 import {TourDocSearchForm} from '../model/forms/tdoc-searchform';
 import {TourDocSearchResult} from '../model/container/tdoc-searchresult';
 import {TourDocAdapterResponseMapper} from './tdoc-adapter-response.mapper';
@@ -6,13 +6,16 @@ import {ItemsJsConfig, ItemsJsSelectQueryData} from '@dps/mycms-commons/dist/sea
 import {GenericItemsJsAdapter, ItemJsResult} from '@dps/mycms-commons/dist/search-commons/services/generic-itemsjs.adapter';
 import {Mapper} from 'js-data';
 import {AdapterQuery} from '@dps/mycms-commons/dist/search-commons/services/mapper.utils';
-import {GenericAdapterResponseMapper} from '@dps/mycms-commons/dist/search-commons/services/generic-adapter-response.mapper';
-import {isNumeric} from 'rxjs/internal-compatibility';
+import {ExtendedItemsJsConfig} from './itemsjs.dataimporter';
 
 export class TourDocItemsJsAdapter extends GenericItemsJsAdapter<TourDocRecord, TourDocSearchForm, TourDocSearchResult> {
-    public static aggregationFields = ['id', 'image_id_i', 'loc_id_i', 'route_id_i', 'destination_id_s', 'track_id_i',
-        'trip_id_i', 'news_id_i', 'video_id_i', 'info_id_i', 'poi_id_i', 'image_similar_id_i', 'loc_parent_id_i'];
-    public static itemsJsConfig: ItemsJsConfig = {
+    public static itemsJsConfig: ExtendedItemsJsConfig = {
+        skipMediaCheck: true,
+        aggregationFields: ['id', 'image_id_i', 'loc_id_i', 'route_id_i', 'destination_id_s', 'track_id_i',
+            'trip_id_i', 'news_id_i', 'video_id_i', 'info_id_i', 'poi_id_i', 'image_similar_id_i', 'loc_parent_id_i'],
+        refConfigs: [
+            { containerField: 'linkedinfos_clob', refField: 'refId', filterFields: ['info_id_i', 'info_id_is']}
+        ],
         spatialField: 'geo_loc_p',
         spatialSortKey: 'distance',
         searchableFields: ['id', 'image_id_i', 'loc_id_i', 'route_id_i', 'destination_id_s', 'track_id_i', 'trip_id_i',
@@ -324,236 +327,9 @@ export class TourDocItemsJsAdapter extends GenericItemsJsAdapter<TourDocRecord, 
         }
     };
 
-    public static createRecordFromJson(responseMapper: GenericAdapterResponseMapper, mapper: Mapper, props: any): any {
-        for (const key in TourDocRecordRelation.hasOne) {
-            if (props[key]) {
-                for (const fieldName in props[key]) {
-                    props[key + '.' + fieldName] = props[key][fieldName];
-                }
-            }
-        }
-
-        for (const key in TourDocRecordRelation.hasMany) {
-            if (props[key]) {
-            }
-        }
-
-        return responseMapper.mapValuesToRecord(mapper, props);
-    }
-
-    public static extendAdapterDocument(values: {}) {
-        // remap fields with fallbacks
-        values['actiontype_s'] = values['actiontype_s'] || values['subtype_s'];
-        values['dateshow_dt'] = values['dateshow_dt'] || values['datestart_dt']
-        values['html'] = values['name_s'] + ' ' +  values['desc_txt'];
-
-        // prepare aggregations
-        for (const filterBase of ['keywords', 'objects', 'persons', 'playlists']) {
-            values[filterBase + '_ss'] = values[filterBase + '_txt'];
-        }
-
-        for (const aggreationName in TourDocItemsJsAdapter.itemsJsConfig.aggregations) {
-            const aggregation = TourDocItemsJsAdapter.itemsJsConfig.aggregations[aggreationName];
-            if (aggregation.filterFunction) {
-                values[aggreationName] = aggregation.filterFunction.call(this, values);
-            } else if (aggregation['mapField']) {
-                values[aggreationName] = values[aggregation['mapField']];
-            }
-        }
-
-        // override some aggregations
-        values['type_txt'] = values['type_txt'] ? values['type_txt'] : values['type_s'];
-        values['type_txt'] = values['type_txt'] ? values['type_txt'].toLowerCase() : '';
-
-        values['year_is'] = values['dateshow_dt']
-            ? new Date(values['dateshow_dt']).getFullYear()
-            : undefined;
-        values['month_is'] = values['dateshow_dt']
-            ? new Date(values['dateshow_dt']).getMonth() + 1
-            : undefined;
-        values['done_ss'] = values['dateshow_dt']
-            ? 'DONE1'
-            : 'DONE0';
-
-        if (values['loc_lochirarchie_s']) {
-            values['loc_lochirarchie_txt'] = values['loc_lochirarchie_s'].split(',,');
-        }
-
-        // add aggregations for searchableFields
-        for (const fieldName of TourDocItemsJsAdapter.itemsJsConfig.searchableFields) {
-            if (fieldName.endsWith('_i') || fieldName.endsWith('_s')) {
-                values[fieldName + 's'] = values[fieldName];
-            }
-        }
-
-        // remap to String because itemjs is string-search-engine ;-)
-        for (const fieldName of [].concat(TourDocItemsJsAdapter.aggregationFields).concat()) {
-            if (isNumeric(values[fieldName])) {
-                values[fieldName] = values[fieldName] + '';
-            }
-        }
-
-        for (const key in values) {
-            if ((key.endsWith('_ss') || key.endsWith('_is') || key.endsWith('_fs')) && isNumeric(values[key])) {
-                values[key] = values[key] + '';
-            }
-        }
-
-        return values;
-    }
-
-    public static lazyCheckForFilePath(pathes: {}, needle: string): string {
-        const normalizedTries = [];
-        if (!needle) {
-            return undefined;
-        }
-
-        normalizedTries.push(needle);
-        if (pathes[needle]) {
-            return needle;
-        }
-
-        let normalized = needle.replace(/^\//, '');
-        normalizedTries.push(normalized);
-        if (pathes[normalized]) {
-            return normalized;
-        }
-
-        normalized = needle.replace('/', '_');
-        normalizedTries.push(normalized);
-        if (pathes[normalized]) {
-            return normalized;
-        }
-
-        normalized = '/' + needle;
-        normalizedTries.push(normalized);
-        if (pathes[normalized]) {
-            return normalized;
-        }
-
-        normalized = '/' + needle.replace('/', '_');
-        normalizedTries.push(normalized);
-        if (pathes[normalized]) {
-            return normalized;
-        }
-
-        if (needle.length === 1) {
-            // console.debug('no matching path found - checked normalized pathes', normalizedTries);
-            return undefined;
-        }
-
-        normalized = needle.substr(0, 1)
-            + needle.substr(1, needle.length).replace('/', '_');
-        normalizedTries.push(normalized);
-        if (pathes[normalized]) {
-            return normalized;
-        }
-
-        // console.debug('no matching path found - checked normalized pathes', normalizedTries);
-
-        return undefined;
-    }
-
-    public static clearNotExistingMediaPathes(records: any[], imagePathes: {}, videoPathes: {}) {
-        // delete reference if media-path not exists in media-container
-        for (const record of records) {
-            let mediaUrl: string = record['i_fav_url_txt'];
-            if (mediaUrl && record['type_s'] !== 'IMAGE' && !imagePathes[mediaUrl]) {
-                const normalizedUrl = TourDocItemsJsAdapter.lazyCheckForFilePath(imagePathes, mediaUrl);
-                if (normalizedUrl && imagePathes[normalizedUrl]) {
-                    // console.debug('FavImage REMAPPED i_fav_url_txt remapped:', mediaUrl, normalizedUrl);
-                    record['i_fav_url_txt'] = normalizedUrl;
-                } else {
-                    // console.debug('FavImage NOT FOUND i_fav_url_txt:', mediaUrl);
-                    delete record['i_fav_url_txt'];
-                }
-            }
-
-            mediaUrl = record['v_fav_url_txt'];
-            if (mediaUrl && record['type_s'] !== 'VIDEO' && !imagePathes[mediaUrl]) {
-                const normalizedUrl = TourDocItemsJsAdapter.lazyCheckForFilePath(videoPathes, mediaUrl);
-                if (normalizedUrl && videoPathes[normalizedUrl]) {
-                    // console.debug('FavVideo REMAPPED v_fav_url_txt remapped:', mediaUrl, normalizedUrl);
-                    record['v_fav_url_txt'] = normalizedUrl;
-                } else {
-                    // console.debug('FavVideo RESET v_fav_url_txt:', mediaUrl);
-                    delete record['v_fav_url_txt'];
-                }
-            }
-        }
-    }
-
-    constructor(config: any, data: any) {
-        for (const aggreationName in TourDocItemsJsAdapter.itemsJsConfig.aggregations) {
-            const aggregation = TourDocItemsJsAdapter.itemsJsConfig.aggregations[aggreationName];
-            if (!aggregation['field']) {
-                aggregation['field'] = aggreationName;
-            }
-        }
-
-        for (const fieldName of TourDocItemsJsAdapter.aggregationFields) {
-            if (fieldName.endsWith('_i') || fieldName.endsWith('_s')) {
-                if (!TourDocItemsJsAdapter.itemsJsConfig.aggregations[fieldName]) {
-                    TourDocItemsJsAdapter.itemsJsConfig.aggregations[fieldName] = {
-                        conjunction: false,
-                        sort: 'term',
-                        order: 'asc',
-                        hide_zero_doc_count: true,
-                        size: 9999
-                    };
-                }
-
-                if (!TourDocItemsJsAdapter.itemsJsConfig.aggregations[fieldName + 's']) {
-                    TourDocItemsJsAdapter.itemsJsConfig.aggregations[fieldName + 's'] = {
-                        conjunction: false,
-                        sort: 'term',
-                        order: 'asc',
-                        hide_zero_doc_count: true,
-                        size: 9999
-                    };
-                }
-            }
-        }
-
-        const recordIds = {};
-        const records = [];
-        const imagePathes = {};
-        const videoPathes = {};
-        // check for duplicates and fill media-container
-        for (const record of data) {
-            if (!record['id'] || record['id'] === '') {
-                console.warn('SKIPPED record - no id', record['id'], record);
-                continue;
-            }
-
-            if (recordIds[record['id']]) {
-                console.warn('SKIPPED record - id already exists id/existing/skipped', record['id'], recordIds[record['id']], record);
-                continue;
-            }
-
-            records.push(record);
-            recordIds[record['id']] = records.length - 1;
-
-            delete record['gpstracks_basefile_s'];
-
-            let mediaUrl: string = record['i_fav_url_txt'];
-            if (mediaUrl && record['type_s'] === 'IMAGE') {
-                imagePathes[mediaUrl] = mediaUrl;
-            }
-
-            mediaUrl = record['v_fav_url_txt'];
-            if (mediaUrl && record['type_s'] === 'VIDEO') {
-                videoPathes[mediaUrl] = mediaUrl;
-            }
-        }
-
-        if (!config['skipMediaCheck']) {
-            TourDocItemsJsAdapter.clearNotExistingMediaPathes(records, imagePathes, videoPathes);
-        }
-
-        console.debug('init itemsjs with config', TourDocItemsJsAdapter.itemsJsConfig, records);
-
-        super(config, new TourDocAdapterResponseMapper(config), records, TourDocItemsJsAdapter.itemsJsConfig);
+    constructor(config: any, records: any, itemsJsConfig: ExtendedItemsJsConfig) {
+        console.debug('init itemsjs with config', itemsJsConfig, records);
+        super(config, new TourDocAdapterResponseMapper(config), records, itemsJsConfig);
     }
 
     mapToAdapterDocument(props: any): any {
@@ -564,6 +340,7 @@ export class TourDocItemsJsAdapter extends GenericItemsJsAdapter<TourDocRecord, 
         return TourDocItemsJsAdapter.itemsJsConfig;
     }
 
+    // TODO move to commons
     protected queryTransformToAdapterQueryWithMethod(method: string, mapper: Mapper, params: any, opts: any): ItemsJsSelectQueryData {
         // map html to fulltext
         const adapterQuery = <AdapterQuery> params;
@@ -598,6 +375,7 @@ export class TourDocItemsJsAdapter extends GenericItemsJsAdapter<TourDocRecord, 
         return queryData;
     }
 
+    // TODO move to commons
     extractRecordsFromRequestResult(mapper: Mapper, result: ItemJsResult): TourDocRecord[] {
         // got documents
         const docs = result.data.items;